+++ source/dark.aua/app.hpp
@@ -96,7 +96,7 @@ namespace apn::dark
 							nullptr));
 						while (::ReadDirectoryChangesW(
 							file.get(),
-							buffer.data(), buffer.size(),
+							buffer.data(), (DWORD)buffer.size(),
 							TRUE,
 							FILE_NOTIFY_CHANGE_FILE_NAME |
 							FILE_NOTIFY_CHANGE_DIR_NAME |
