+++ source/text_drop.aua/utils.hpp
@@ -23,7 +23,7 @@ namespace apn::text_drop
 		ofn.Flags = flags;
 		ofn.lpstrTitle = title;
 		ofn.lpstrFile = file_name.data();
-		ofn.nMaxFile = file_name.size();
+		ofn.nMaxFile = (DWORD)file_name.size();
 		ofn.lpstrFilter = filter;
 
 		// ファイル選択ダイアログを表示します。
@@ -57,7 +57,7 @@ namespace apn::text_drop
 		ofn.Flags = flags;
 		ofn.lpstrTitle = title;
 		ofn.lpstrFile = file_name.data();
-		ofn.nMaxFile = file_name.size();
+		ofn.nMaxFile = (DWORD)file_name.size();
 		ofn.lpstrFilter = filter;
 		ofn.lpstrDefExt = default_extension;
