+++ source/workspace.aua/dock/layout_list.hpp
@@ -49,7 +49,7 @@ namespace apn::workspace
 						nullptr));
 					while (::ReadDirectoryChangesW(
 						file.get(),
-						buffer.data(), buffer.size(),
+						buffer.data(), (DWORD)buffer.size(),
 						TRUE,
 						FILE_NOTIFY_CHANGE_FILE_NAME |
 						FILE_NOTIFY_CHANGE_DIR_NAME |
